fix(trusty-mpm): sessions/project list shows daemon-managed sessions (#3822)
`ProjectRegistry::auto_register_from_sessions` only ever ran ONCE per
daemon process -- inside `DaemonState::project_registry()`'s
`OnceCell::get_or_init`, seeded from whatever session snapshot existed
the moment ANY of its 15+ call sites (a status poll, an MCP tool,
`resolve_gh_env`, ...) first touched the registry, frequently before
any session existed. Nothing else called that pass again, so a
project implied by a session spawned afterward was registered
nowhere -- even though the daemon plainly knew about the session
(`tm stop`/`tm resume` by explicit name still worked, since those go
through the separate session-manager store, not the project
registry). Result: `tm project list` reported "no projects
registered" for a live, daemon-managed session forever.

Add `ProjectRegistry::register_from_session`, an explicit per-session
counterpart to the one-shot boot pass, called at spawn time from all
three managed-session creation paths (`spawn_managed_cloned`,
`spawn_managed_inproject`, `spawn_managed_local` in
`daemon::managed_routes::lifecycle`) right after the session record
is persisted -- registration no longer depends on which caller
happened to touch the registry first.

Also: `tm sessions list` and `tm sessions ls` turned out to be two
DIFFERENT, intentionally-scoped commands (`list` = this project's
local pre-session-manager sessions; `ls` = daemon-managed sessions
from anywhere, already correctly global with no scoping bug) -- not
a second instance of the same bug, just command-name confusion that
produced the "no sessions" half of the original report. Added a
one-line discoverability nudge pointing an empty `list` result at
`ls`.

Closes #3822
